12 - refactor: moved database functions to model
