+++ controller/index.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/model"
-	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/shared/db"
 	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/shared/session"
 	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/shared/view"
 	"log"
@@ -28,7 +27,7 @@ func IndexGET(w http.ResponseWriter, r *http.Request) {
 	v := view.New(r)
 	v.Name = "index"
 	v.Vars["Auth"] = user.Authenticated
-	v.Vars["Courses"] = db.GetCoursesToUser(user.ID)
+	v.Vars["Courses"] = model.GetCoursesToUser(user.ID)
 	v.Vars["Message"] = joinedCourse
 	v.Render(w)
 }
@@ -38,7 +37,7 @@ func JoinCoursePOST(w http.ResponseWriter, r *http.Request) {
 	joinedCourse = ""
 
 	// Check if course exists
-	course := db.CourseExists(r.FormValue("courseID"))
+	course := model.CourseExists(r.FormValue("courseID"))
 
 	// If course ID == "", it doesn't exist
 	if course.ID == -1 {
@@ -50,21 +49,21 @@ func JoinCoursePOST(w http.ResponseWriter, r *http.Request) {
 	user := session.GetUserFromSession(r)
 
 	// Check that user isn't in this class
-	if db.UserExistsInCourse(user.ID, course.ID) {
+	if model.UserExistsInCourse(user.ID, course.ID) {
 		//joinedCourse = ""
 		ErrorHandler(w, r, http.StatusBadRequest)
 		return
 	}
 
 	// Add user to course if possible
-	if !db.AddUserToCourse(user.ID, course.ID) {
+	if !model.AddUserToCourse(user.ID, course.ID) {
 		ErrorHandler(w, r, http.StatusBadRequest)
 		return
 	}
 
 	// Log joinedCourse in the database and give error if something went wrong
 	lodData := model.Log{UserID: user.ID, Activity: model.JoinedCourse, CourseID: course.ID}
-	if !db.LogToDB(lodData) {
+	if !model.LogToDB(lodData) {
 		log.Fatal("Could not save JoinCourse log to database! (index.go)")
 	}
